Make things constant
The color information for red and green never change. Rather than
handling them as callables, they can be calculated ahead of time,
calling out to `curses.color_pair` only once for each color.

Also, the unicode string for the arrow is the same everywhere its used.
`ARROW` is being created as a reference to it. Now anything that needs
to print the symbol can use this instead.